Removes Python2 from setup and test scripts
Why is this change needed?
--------------------------
Python2 is no longer being maintained as of 2020. Since this is the case
we will not be supporting Python2 in the next version of the DocRaptor
agent.

How does it address the issue?
------------------------------
This removes the installation steps for Python2 from script/setup. It
also updates script/test to only test against Python3.

Co-authored-by: David Ronk <david@expectedbehavior.com>