Updates swagger configuration for document strict parameter
Why is this change needed?
--------------------------
The strict parameter only accepts 2 specific strings and it is not
required to make a document. This specific configuration was causing
some weird behavior in our C# agent so we changed it there. In an effort
to keep configurations consistent we should update it here as well.

How does it address the issue?
------------------------------
This updates the configuration in the following ways:
- It removes the default value from the strict parameter.
- It specifies both "none" and "html" as a possible value as per the
  DocRaptor documentation.

In addition to the above changes, this reorders the Doc model's
parameters to have the required attributes first, followed by those that
are optional. Many languages suggest that required parameters should be
first. As a result of the configuration change the Doc constructor
parameter order also changes to suit best practices.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://docraptor.com/documentation/api#api_strict