Removes duplicate DocApi configuration that was breaking tests
Why is this change needed?
--------------------------
When the tests were updated to use the new configuration syntax there
were a few places where the old configuration was not fully removed.
This caused a few of the test files to error.

How does it address the issue?
------------------------------
Removing the duplicate DocApi configuration fixes the tests.