Update example error handling to use the correct error attributes
Why is this change needed?
--------------------------
When an example encounters an API exception we should output helpful
details about what went wrong.

How does it address the issue?
------------------------------
When we rescue from an API exception we will now print out helpful
details about the status, reason, and body of the error response.