Change tests to use test api key
Why is this change needed?
--------------------------
It was confusing that the test api key wasn't valid for testing hosted
documents.  It also complicated the tests since they required a valid
paid api key to pass.

How does it address the issue?
------------------------------
DocRaptor has been updated to allow hosted document creation using the
test api key.  This updates the agent to do the same.