Move generator name off the command line
Why is this change needed?
--------------------------
For python the generator is `python-legacy`, and it's annoying to have
a different command for one of the languages.

We use `python-legacy` instead of `python` because the newer clients
work dramatically differently, in a more verbose and less intuitive way.
There is some discussion in GitHub about the next python generator being
closer to the other languages in style, which would be nice. Since it
looks like there are more changes on the horizon we'll wait to make
changes that'll require user code changes.

How does it address the issue?
------------------------------
Move the generator name to a file so the script can match between repos
but the correct client will be generated.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?