Allow username or password without the other
Why is this change needed?
--------------------------
This prevents requiring code changes for users. Openapi changed to
require both for basic auth, but since we use an API token we don't need
both.

How does it address the issue?
------------------------------
Change this conditional to be an `or` instead of `and`.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?