Add/fix bytes type responses
Why is this change needed?
--------------------------
The `python-legacy` generator doesn't supports the bytes type response.

I believe originally swagger used the `binary` response type to indicate
a binary string. Later this changed so `binary` returned a file
reference, and `bytes` was added to allow returning a binary string.

Previously we fixed this with a patch[1], but it wasn't obvious when
debugging issues with the newly generated client. First I went down the
path of adding `bytes` support. Then I discovered our previous patch.

How does it address the issue?
------------------------------
Since I had already done the work to add `bytes` support I updated the
patch to work with it so we could use the same `docraptor.yaml` file
in python as the other languages (`bytes` instead of `binary`).

I also added a more obvious method of client patching to hopefully avoid
lost time in the future. Having a patches directory should make it clear
what modifications we make to the autogenerated code and the related
commits should make the context much more accessible.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
[1] - 0cd0b7a7af7ae72f37dd0ce825b1903e98824bd0
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?