Update tmp gitignore
Why is this change needed?
--------------------------
We want to know when the generator changes the top level .gitignore. The
fewer changes to that file the easier it is to review those changes.

How does it address the issue?
------------------------------
Move ignoring into the tmp directory. Also we can ignore everything in
there since it's tmp, instead of just ignoring test_output.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?