Remove .gitignore customizations
Why is this change needed?
--------------------------
We'd like to see future generated modifications so they're easy to add.
We don't want those differences cluttered by our customizations.

How does it address the issue?
------------------------------
Remove our customizations. It's fine to have the autogenerated files
committed, which makes it easier to review changes and check if we want
to change our processes based on those changes. For instance, we might
want to start using travis.

The other files can be handled elsewhere. .DS_Store should be ignored in
`~/.gitignore`. `.docraptor_key` isn't currently used in this repo, but
if it's added we should put it in `tmp` which will ignore it
automatically.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?