Actually generate code with openapi, java 3.4+
Why is this change needed?
--------------------------
Swagger isn't being maintained. We needed to switch to openapi for
other agents, and if it works it's nice for all the agents to use the
same tech.

How does it address the issue?
------------------------------
This is just the result of `script/generate_language`.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?