Improve generated file handling
Why is this change needed?
--------------------------
See the new `.review/README.md` for the explanation.

How does it address the issue?
------------------------------
Setup a process for reviewing changes to generated content that we only
want to review when it changes.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29124557/todos/5284412189
https://3.basecamp.com/3093825/buckets/29124557/todos/5284421778

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?