Switch to alpine for testing
Why is this change needed?
--------------------------
The oldest debian image had errors, and a quick google indicated the
issue was EOL and lack of support.

How does it address the issue?
------------------------------
Switch to the alpine images, which are what the openapi-generator gitlab
file uses, so presumably the generated code is tested against them
regularly.

This also adds a script/inside_container/init file, which is run
before the command in the docker container.  This is used to install
bash, which is needed for the test script, and is not installed by
default in the alpine images.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://3.basecamp.com/3093825/buckets/29007795/todolists/5257298600

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?